Add facility to plug-in custom DB connection pools
Dropwizard by default is shipped with the Tomcat JDBC Connection pool.
It's mostly fine, but has some disadvantages (performance, limited metrics,
handling network failures). It would be great if we had facility to
plug-in other DB pools into Dropwizard. For example, Brett Wooldridge's
wonderful HikariCP or BoneCP.

Of cource, nothing stops from implementing a factory, that creates a
ManagedDataSource. The problem is that dropwizard-jdbi, dropwizard-hibernate
and dropwizard-migrates modules require DataSourceFactory for the configuration.

DataSourceFactory is closely tied with the Tomcat pool. So if we use another
pool, we are forced to write own factories for JDBI, Hibernate, etc...

A solution is to provide an interface that abstracts a configuration of
a connection pool. Database manipulation modules will be linked with the
interface, but not with DataSourceFactory. A custom pool factory just need to
implement this interface, and it could be plug-in as a Tomcat pool
replacement.

If we writed Dropwizard from the scratch we would move the Tomcat pool
to a separate module from dropwizard-db. But now we have many users and it
would be a nightmare for them to migrate to a new implementation.

So we are forced to retain the Tomcat JDBC pool in dropwizard-db. New pool
users should be advised to exclude it manually in the Maven configuration.