Merge dropwizard-java8 addon into mainline
dropwizard-java8 used to provide support for Java 8 classes like `Optional<T>` and
the Java Date and Time API (JSR-310) for Dropwizard. Given that the baseline for
Dropwizard is now Java 8, those additions can be maintained in the mainline repository.

The only backward-incompatible change is in the `io.dropwizard.auth.Authenticator` interface
which is now using `java.util.Optional` instead of the equivalent class from Google Guava.

Closes #962